Fluxの構成要素をcommonパッケージに移動 ・Action ・Dispatcher ・Store
