+++ app/src/main/java/com/example/weatheralarmapp/flux/common/Action.kt
@@ -1,0 +1,5 @@
+package com.example.weatheralarmapp.flux.common
+
+interface Action {
+    val payload: Any?
+}
